stock-analyzer/pyproject.toml
[project]
name = "stock-analyzer"
version = "0.1.0"
description = "A powerful stock analysis tool built with Google ADK that provides comprehensive investment recommendations by analyzing news, historical performance, economic factors, and political risks."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.18",
    "duckduckgo-search>=8.0.2",
    "google-adk>=0.5.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "litellm>=1.70.0",
    "python-dotenv>=1.1.0",
]
